meson: Add nlohmann_json dep
In the future, we will need to parse some JSON. I've already worked with
nlohmann's library, so I think it's a good choice.

Also, try to use the meson wrap feature to not depend on system
libraries availability. If the lib is not found, it will download and
compile the sources automagically.