cmd: Add `track` command
We need a command for tracking the current price of a stock and send a
email alerts if the price crosses the sell or buy price.

For that it gets the email information provided by the configuration
file and tracks the price using the mfinace API every 15 minutes.

This command needs some new configuration options, describe them on the
README and add them to the config example.