cmd: Create a way to have subcmds
Our application will be self discoverable, meaning that the user can
discover the application features by just interactively running the
application and seeing the help messages.

With that in mind, splitting each feature into its own sub-command is a
good idea. This way we can list each command on the help message, and
every sub command has its own help message too.

For that, create the `Cmd` struct. This struct is used to represent a
command, it can have subcommands. The goal of this struct is to take
away the work of parsing the args and generating the help message when
creating a new command. The developer will only need to state the needed
arguments and the struct will do the work.

For now, just create a dumb `main_cmd` to serve as the base command that
will contain the useful sub commands in the future.